Start modularizing the client code.
This is the first step toward making the client code not be a terrible
ball of mud. Before, the client code was all in one file because it
needed to initialize some stuff before requiring external modules could
work. Now, a dummy global._hmr[key].initModule() function is placed so
modules can be loaded before HMR is initialized. That does mean that any
modules loaded at this point aren't hot-replaceable but that doesn't
seem like a real issue because I don't know who expects to hot-replace
the HMR internals themselves.